requirements.txt
pdfplumber>=0.9.0
python-dateutil>=2.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
pytest>=7.4.0

